requirements.txt
# AWS
boto3>=1.28.0
botocore>=1.31.0

# Data Processing
pandas>=2.0.0
numpy>=1.24.0
pyarrow>=14.0.0
yfinance>=0.2.35
pyspark>=3.5.0
requests>=2.31.0

# Configuration and CLI
PyYAML>=6.0
python-dotenv>=1.0.0
argparse>=1.4.0

# Data Storage and Formats
fastparquet>=2023.10.0
tables>=3.8.0

# Monitoring
grafana-api>=1.0.3

# Visualization
matplotlib>=3.7.0
seaborn>=0.13.0

# Utilities
tqdm>=4.65.0